fix: actualización de src
