--- Persona.java
@@ -1,36 +1,0 @@
-/*
- * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
- * Click nbfs://nbhost/SystemFileSystem/Templates/Classes/Class.java to edit this template
- */
-package repositorioremotogrupo5;
-
-/**
- *
- * @author Lin Reyes
- */
-public abstract class Persona {
-    
-    public String nombre;
-    public String apellido;
-    public int edad;
-    public double peso;
-
-    public void setNombre(String nombre) {
-        this.nombre = nombre;
-    }
-
-    public void setApellido(String apellido) {
-        this.apellido = apellido;
-    }
-
-    public void setEdad(int edad) {
-        this.edad = edad;
-    }
-
-    public void setPeso(double peso) {
-        this.peso = peso;
-    }
-    
-    public abstract void ocupacion();
-    
-}
